:recycle: move common otlp code from api to internal/utils/otlp
